fix(core): SMA-312 use u64 for SessionEvent::Compacted.original_count
usize is platform-dependent (32-bit on i686/wasm32, 64-bit on x86_64),
which means a Compacted event written on a 64-bit producer could
silently truncate when read on a 32-bit consumer. Since SessionEvent is
serialized into the persisted event log, the field must be a
fixed-width integer.

Caught by CodeRabbit on PR #17.

+++ crates/paigasus-helikon-core/src/session.rs
@@ -103,8 +103,10 @@ pub enum SessionEvent {
     Compacted {
         /// LLM-produced summary.
         summary: String,
-        /// Number of events the summary replaces.
-        original_count: usize,
+        /// Number of events the summary replaces. `u64` (not `usize`) because
+        /// the value is serialized into the persisted event log — a 32-bit
+        /// consumer must read what a 64-bit producer wrote without truncation.
+        original_count: u64,
     },
 }
